Issue #1652 Add a real widget displaying the length of the line in the panel: correction to open already save studies.

+++ src/SketchPlugin/SketchPlugin_Validators.cpp
@@ -903,7 +903,7 @@ bool SketchPlugin_SplitValidator::isValid(const AttributePtr& theAttribute,
     ModelGeomAlgo_Point2D::getPointsInsideShape(anAttrShape, aRefAttributes, aC->pnt(),
                                                 aX->dir(), aDirY, aPoints);
 
-    int aCoincidentToFeature = aPoints.size();
+    int aCoincidentToFeature = (int)aPoints.size();
     if (aKind == SketchPlugin_Circle::ID())
       aValid = aCoincidentToFeature >= 2;
     else
